ingest: Remove parameters related to trigger
Realized through https://github.com/nextstrain/rsv/pull/37 that the
ingest pipeline does _not_ trigger the rebuild. The rebuild is just
scheduled to run after the ingest workflow. Removing all parameters and
references to trigger in this commit so that it does not confuse anyone
else in the future.

Keeping the schedule as-is since it's been working fine and we are
planning to be shift pathogen workflows in the future to be able to
go from ingest to a build within a single run without going through
triggers and S3 interactions.