Allows for CDS (as well as gene) features to generate a new gene reference
The newreference.py script processes a GenBank file, producing new gene
reference files (GenBank and FASTA) which are required for creating the
Nextstrain gene tree views.

As we intend to use this script as a template for other viruses (e.g. dengue, measles),
it was necessary to allow the use of CDS features in the GenBank file, not just the gene.