[ingest] Embed URLs in curated data
This will make it simpler for phylo workflows (see next commit) and
co-locates the URLs with the data fields in the intermediate files which
should better adhere to data-use policies.

Note: There's a number of ways we could encode the URLs:
1. In the field as `VALUE: URL`
2. In the field as `VALUE <URL>`
3. As a separate column, e.g. 'ppx_accession' and 'ppx_accession_url'

This commit chooses 1. Related Augur PR <https://github.com/nextstrain/augur/pull/1852>
can interpret (1) and (2).