[phylo] use URLs from metadata fields for export
This leverages the work done in the previous commit to encode URLs within
the associated values. This only works with the (currently unmerged)
Augur PR <https://github.com/nextstrain/augur/pull/1852>. The benefit
is that phylo workflows don't have to special case anything, it's all
handled on the ingest side.