Add multiple input support
Updated the filepaths for the rules copied from pathogen-repo-guide 
to support the `a_or_b` wildcard. It's not clear to me why the workflow uses 
`a_or_b` instead of a `subtype` wildcard to match the config param `subtypes`, 
but I'm not going to make the changes to consolidate them here.

Removes snakemake_rules/download.smk since the multiple input support uses 
the Snakemake storage plugins to handle remote files. This bumps the minimum 
Snakemake version to 8.0.0.