ingest: Add rules to extract "RESTRICTED" data
Extract "OPEN" and "RESTRICTED" data into separate files that are uploaded to 
S3 separately. This will reduce the amount of duplicate data that we host on S3.

Outside of the changes in the workflow, we should delete the previously uploaded
"*_with_restricted" files from S3 so that they are not confused with the new 
"*_restricted" files added here.