.github/workflows/fetch-and-ingest: use docker runtime
I noticed in a recent run of the ingest workflow that ~6min was spent on waiting 
for the Batch job to start. So I was curious if this workflow can run within
GH Actions using the docker runtime. My trial run was 9min23s¹ compared to the 
previous run on AWS Batch that was 15m21s.² 

Just going to use the docker runtime since it's faster and free

¹ <https://github.com/nextstrain/rsv/actions/runs/18957867813>
² <https://github.com/nextstrain/rsv/actions/runs/18881097661>